Initiate the kube-apiserver audit log permissions on start
While the directory permissions are limited on starting the
openshift-kube-apiserver pod; this also initiates the audit logs
permissions to a more secure 0600.

This makes the underlying library (lumberjack) reuse these permissions
throughout the deployment, instead of keeping using the 0644 which was
the default.

Note that the file permissions in the directory were already providing
protection from accessing those files. This merely promotes good
patterns for the log files.