Report infra provider and feature set metrics
The infrastructure provider metric summarizes the desired type and
region (if any) to prometheus, based on the desired config. If type
is not recognized it is passed along, if it is empty it is set as
empty.

The featureset metric reports what featureset is in use and 1 if it
is "supported" and 0 if it is "not supported" (in the tech preview
sense).