add privilege to operator init container
CRI-O 1.16 made a change to how selinux labels were handled that increased security. In doing so, it tightened security and needs to be explicitly told to not label a container. In the case of this initContainer, it is trying to edit the permissions of a file on the host system, which is (to SELinux) a security flaw, and thus selinux is denying it.

explicitly add privilege to the initContainer so it can touch the hostPath

Signed-off-by: Peter Hunt <pehunt@redhat.com>