baremetal: Explicitly advertise the host IP chosen by kubelet
We've updated MCO to ensure IPv6 VIPs are not accidentally used by
kubelet.  Make use of that value and set it as the address
kube-apiserver should advertise.  Otherwise, it uses the same broken
default behavior that kubelet does and will incorrectly choose IPv6
VIPs, breaking the cluster.

The change to specify the correct IP to kubelet is:
https://github.com/openshift/machine-config-operator/pull/1444