kube-apiserver-cert-syncer-kubeconfig: point to serving cert
The service account gets the service-account CA that is valid at the point of creation. So if the localhost-recovery CA is not yet part of the kube-apiserver serving CA (which is the case during bootstrapping), this service-account won't work with the localhost-recovery endpoints.

The serving CA bundle though is always correct as it changes with time.

Fixes https://github.com/openshift/cluster-kube-apiserver-operator/pull/663/commits/2a44f96452f248818ff4c2467d51eef50fd139f2#r456357777