render: read service-account-issuer from authentication config file
ref: https://issues.redhat.com/browse/CO-1266

We need to support user-provided service-account-issuer for supporting user's pre existing IAM OIDC provider.
The user will provider the authentications.config.openshift.io/cluster object with the issuer at install-time
and the bootstrap-kube-apiserver should use that issuer to prevent in divergence from the real kube-apiserver.

So the render command accepts a flag pointing to a file for authentication object, which the render command reads
to extract the issuer from the spec and updates the bootstrap-apiserver-config.