render: accept user provided bound-service-account-signing keys for kube-apiserver
To support pre-existing IAM OIDC provider on AWS, we need an option to configure the kube-apiserver with user-provided keys for service-account-signing-keys.
This enables that the tokens issued by the cluster's kube-apiserver will be trusted by AWS IAM as setup by the OIDC provider. see https://issues.redhat.com/browse/CO-1266

When the render sees that the user has provided `bound-service-account-signing-key.{key,pub}`, the renderer configures the
bootstrap kube-apiserver with that specific service-account-signing-key and also creates the `next-bound-service-account-signing-key` secret
which will be pushed to the cluster and used by the kube-apiserver-operator to re-use those user provided keys for in-cluster kube-apiserver.

The NewBoundSATokenSignerController [1] already suports adopting the secret when the keys are already set, and this allows the kap-o to create all the rest of the
public ConfigMap and operand Secret from this user-provided secret.

[1]: https://github.com/openshift/cluster-kube-apiserver-operator/blob/0cf5528e6410ed723ab9be701d8e9f3e7a50627b/pkg/operator/boundsatokensignercontroller/controller.go#L56-L57