bindata: kube-apiservers overstate steady-state slightly
The kube-apiserver is one of our most dynamic scaling components,
but the majority of clusters have a base load that is fairly
predictable of around 300 millicore. In order to get more headroom
for smaller clusters, adjust the out of the box request to a more
accurate default. All of the sidecars were measured in heavily
loaded production clusters as taking less than 5 millicore each,
and the core kube-apiserver is tweaked slightly.  CPU is burstable
so the apiserver will grow as necessary into slack capacity, while
still yielding to etcd on saturated masters.