prometheus rules for client-go
Kubernetes is a distributed system, so network is a fundamental part
of the system.

At the heart of the system we can find the apiservers and the etcd
cluster, both have a good instrumentation that allow users and
developers to diagnose and troubleshoot networking problems.

However, there are other components like the kube-scheduler,
kube-controller-manager and kubelets that depend on its connectivity
against the apiserver or aggregated apiservers to work.

These communication can be achieved via a direct network connection,
and intermeadiate load balancer or via the Kubernetes Services
implementation (CNI, ...)

Recording the latency and the error rate of these components against
the apiserver will allow users and developers to understand and
troubleshoot the networking problem caused by external factors.

Also, before adding SLOs we need a better understanding on how these
SLIs behave, since they depend a lot of the environment, per example,
cloud providers load balancers vs BareMetal keepalived+haproxy or
customer load balancer, ...

Co-authored-by: Damien Grisonnet <damien.grisonnet@epita.fr>