Allow ephemeral volumes in all SCCs
"ephemeral" volumes, called as Generic Ephemeral Volumes in upstream docs,
are volumes represented by a PersistentVolumeClaim template in Pod.

Using these volumes is safe for any user, as users with the capability to
use "ephemeral" volumes can fool Kubernetes to only create PVCs, in the
same fashion as users that can create StatefulSets can fool Kubernetes to
create PVCs too.

In both cases, the created PVC is subject of namespace's PVC quota and it
may be dynamically provisioned using a StorageClass, just as a regular PVC.

Users won't get ability to create PVs this way nor ability to refer to an
existing volume.