manifests: fix the scope of TechPreviewNoUpgrade
The TechPreviewNoUpgrade alert used to have an expression that was only
excluding the default feature set meaning that all other feature set
were reported as causing the cluster to be unupgradeable. However, that
is not the case for the LatencySensitive feature set. As such, this
change limit the list of feature set causing this alert to fire to:

  - TechPreviewNoUpgrade
  - CustomNoUpgrade

Also, since the original name of this alert wasn't generic enough
compare to its scope, it was renamed to UnsupportedFeatureSetEnabled.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>