alerts: exclude iowait and steal CPU modes
Exclude the iowait and steal CPU modes on top of the idle mode since
these correspond to idle/wait states.

Per the iostat man page:

%idle
Show the percentage of time that the CPU or CPUs were idle and the
system did not have an outstanding disk I/O request.

%iowait
Show the percentage of time that the CPU or CPUs were idle during
which the system had an outstanding disk I/O request.

%steal
Show the percentage of time spent in involuntary wait by the
virtual CPU or CPUs while the hypervisor was servicing another
virtual processor.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>