alerts: remove noisy warning alert
The ExtremelyHighIndividualControlPlaneCPU had two different variation
with different criticality.
- The warning alert was firing after 5 minutes of the control plane CPU
  usage being above 90%
- The critical alert firing after 60 minutes of the control plane CPU
  usage being above 90%

Because the timeline of a warning level alert in OCP was defined to be
60 minutes, having the warning level alert response tiemline collide
with the time we expect response from the critical alert makes it noisy.

On top of that, 5 minutes is very low considering that this alert most
likely always fire during upgrades because of the pressure control plane
nodes are suffering from post upgrade.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>