manifests: rename API performance dashboard
In previous PR this manifests was labelled as "available only when Console capability enabled". This causes CVO to force enable Console capability when upgrading from baseline 4.13 cluster - as this manifest is present.

In order to avoid this, the manifest needs to be renamed, so that CVO would treat it as a new one (since its applicability has changed)