use AlwaysAllow UnhealthyPodEvictionPolicy option
allow eviction of unhealthy (not ready) pods even if there are no disruptions
allowed on a PodDisruptionBudget. This can help to drain/maintain a node and
recover without a manual intervention when multiple instances of nodes or pods
are misbehaving. Use this with caution, as this option can disrupt perspective
pods that have not yet had a chance to become healthy.