Disable PersistentVolumeLabel admission plugin
PersistentVolumeLabel admission plugin needs a valid cloud config.
kube-apiserver does not get it in OCP 4.17, and thus it reject valid
in-tree GCE PersistentVolumes.

The admission plugin was deprecated upstream a long time ago, it was
removed from Kubernetes 1.31 (OCP 4.18) and it cannot work without
a cloud provider in kube-apiserver anyway.