bindata/alerts/slo: fix burnrate calculation
Calculate the request burn rate based on the total number of read+write
requests instead of separately calculating the burn rate for each
request type. This used to cause an erroneous result when summing up the
read and write burn rates together as it wouldn't account for the
propertion of failures amongst all requests.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>