SCC: add hostmount-anyuid-v2
the intent of hostmount-anyuid is to allow a pod access to paths on the host.
The problem is, hostPath volumes aren't selinux relabeled, and there
are paths that the default selinux type `container_t` cannot access.

This breaks expectation, and makes pods that rely on hostmount-anyuid brittle to
selinux changes on the host (see https://issues.redhat.com/browse/OCPBUGS-55013)

Instead of relaxing permissions of all paths on the host, or increasing the
ability of container_t, we should trust pods that are granted access to this already powerful
SCC to use its power fully.

Signed-off-by: Peter Hunt <pehunt@redhat.com>