Use scrapeClass for service monitors
As documented in [1], the platform Prometheus servers are configured
with a scrape class named `tls-client-certificate-auth` which
automatically injects the client certificate and key files for mutual
TLS into the scrape configuration of the service.

This commit also removes the `namespaceSelector` field of the operator's
service monitor: it isn't required when the service monitor and service
are located in the same namespace. It also adjusts the `jobLabel` value
to refer to the `app` label of the service.

[1] https://rhobs-handbook.netlify.app/products/openshiftmonitoring/collecting_metrics.md/#configuring-prometheus-to-scrape-metrics

Signed-off-by: Simon Pasquier <spasquie@redhat.com>